Refactor GUI to use common modules configure_a and visuals.
